apps/backend/requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
sqlalchemy==2.0.25
asyncpg==0.29.0
psycopg2-binary==2.9.9
pydantic==2.5.3
pydantic-settings==2.1.0
python-dotenv==1.0.0
python-jose[cryptography]==3.3.0
bcrypt>=4.1.0
python-multipart==0.0.6
openai==1.7.0
langchain==0.1.0
langchain-openai==0.0.2
celery==5.3.4
redis==5.0.1
httpx==0.26.0
fpdf2==2.8.2
